@@ -26,22 +26,22 @@
     "es5.js"
   ],
   "dependencies": {
-    "eslint-plugin-import": "^1.7.0"
+    "eslint-plugin-import": "^2.7.0"
   },
   "devDependencies": {
-    "babel-eslint": "^7.0.0",
-    "eslint": "^3.0.0",
-    "eslint-find-rules": "^1.14.3",
-    "eslint-plugin-ava": "^3.0.0",
-    "eslint-plugin-flowtype": "^2.31.0",
-    "eslint-plugin-import": "^2.2.0",
-    "eslint-plugin-react": "^6.10.3",
-    "flow-bin": "^0.33.0",
-    "husky": "^0.13.2",
-    "lint-staged": "^3.4.0",
-    "npm-run-all": "^3.1.0",
-    "npmpub": "^3.0.3",
-    "prettier": "^0.22.0"
+    "babel-eslint": "^7.2.3",
+    "eslint": "^4.6.1",
+    "eslint-find-rules": "^3.1.1",
+    "eslint-plugin-ava": "^4.2.2",
+    "eslint-plugin-flowtype": "^2.35.1",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-react": "^7.3.0",
+    "flow-bin": "^0.54.0",
+    "husky": "^0.14.3",
+    "lint-staged": "^4.1.0",
+    "npm-run-all": "^4.1.1",
+    "npmpub": "^3.1.0",
+    "prettier": "^1.6.1"
   },
   "scripts": {
     "#precommit": "setup by husky",
