@@ -1,5 +1,6 @@
 /* eslint-disable import/no-unresolved, import/no-extraneous-dependencies */
-import React, { PropTypes } from "react";
+import React from "react";
+import PropTypes from "prop-types";
 import cx from "classnames";
 
 import styles from "./index.css";
